+++ epan/dissectors/packet-pvfs2.c
@@ -3070,7 +3070,7 @@ pvfs2_io_tracking_hash(gconstpointer k)
 {
 	const pvfs2_io_tracking_key_t *key = (const pvfs2_io_tracking_key_t *) k;
 
-	return (key->tag >> 32) ^ ((guint32) key->tag);
+	return (guint) ((key->tag >> 32) ^ ((guint32) key->tag));
 }
 
 static void
